+++ src/main/java/com/spring/gogidang/service/EventService.java
@@ -1,6 +1,7 @@
 package com.spring.gogidang.service;
 
 import java.util.ArrayList;
+import java.util.List;
 
 import com.spring.gogidang.domain.EventVO;
 
@@ -11,5 +12,7 @@ public interface EventService {
 	public EventVO getByNum(int event_num);
 	public void modifyByNum(int event_num);
 	public void deleteByNum(int event_num);
-	public void register(EventVO eventVO);
+	public int register(EventVO eventVO);
+	
+	public List<EventVO> getList();
 }
