+++ src/main/java/com/spring/gogidang/service/QnaService.java
@@ -8,10 +8,6 @@ import javax.crypto.Cipher;
 import com.spring.gogidang.domain.Criteria;
 import com.spring.gogidang.domain.QnaVO;
 
-
-
-
-
 public interface QnaService {
 	
 	public int getListCount(); // 글 개수 구하기 
@@ -22,4 +18,5 @@ public interface QnaService {
 	public QnaVO qnaModifyForm(int qna_num);
 	public int qnaDelete(HashMap<String, String> hashmap);
 	public int qnaReply(QnaVO qna);
+	public List<QnaVO> getList();
 }
