+++ src/main/java/com/spring/mapper/EventMapper.java
@@ -1,6 +1,7 @@
 package com.spring.mapper;
 
 import java.util.ArrayList;
+import java.util.List;
 
 import com.spring.gogidang.domain.EventVO;
 
@@ -12,5 +13,7 @@ public interface EventMapper {
 	public EventVO getByNum(int event_num);
 	public void modifyByNum(int event_num);
 	public void deleteByNum(int event_num);
-	public void register(EventVO eventVO);
+	public int register(EventVO eventVO);
+	
+	public List<EventVO> getList();
 }
