+++ src/main/resources/com/spring/mapper/NoticeMapper.xml
@@ -21,7 +21,7 @@
 		<selectKey keyProperty="notice_num" order="BEFORE" resultType="int">
 			SELECT nvl(max(notice_num), 0)+1 FROM notice
 		</selectKey>
-		INSERT INTO notice VALUES (#{notice_num}, #{u_id}, #{title}, #{content},sysdate,#{notice_num})
+		INSERT INTO notice VALUES (#{notice_num}, #{u_id}, #{title}, #{content}, sysdate, #{notice_num})
 	</insert>
 	
 	<select id="getDetail" parameterType="int" resultType="NoticeVO">
@@ -38,5 +38,9 @@
 	
 	<select id="isBoardWriter" parameterType="HashMap" resultType="int">
 		SELECT count(*) FROM notice WHERE notice_num=#{notice_num} and u_id=#{u_id}
+	</select>
+	
+	<select id="getList" resultType="NoticeVO">
+		select * from notice
 	</select>
  </mapper>
