+++ src/main/webapp/WEB-INF/views/member/updateList.jsp
@@ -28,9 +28,9 @@
 			<h4>관리자 페이지</h4>
 			<ul>
 				<li><a href="storeWait.st">대기중인 가게 승인</a></li>
-				<li><a href="noticelist.no">공지사항 관리</a></li>
-				<li><a href="qnalist.qn">문의 관리</a></li>
-				<li><a href="eventList.ev">이벤트 관리</a></li>
+				<li><a href="noticeAdmin.no">공지사항 관리</a></li>
+				<li><a href="qnaAdmin.qn">문의 관리</a></li>
+				<li><a href="eventAdmin.ev">이벤트 관리</a></li>
 			</ul>
 			<%
 				} else {
